requirements.txt
streamlit
google-generativeai
langchain-google-genai
faiss-cpu
datasets
Pillow
pytesseract
python-dotenv
pandas
numpy
requests
